feat/style - turbo + fix - ready to test
