+++ assets/styles/_splitter.scss
@@ -261,6 +261,4 @@
     text-decoration: underline;
     color: var(--color-accent);
   }
-
-
 }
