+++ composer.json
@@ -48,6 +48,7 @@
     "symfony/uid": "7.0.*",
     "symfony/ux-icons": "^2.24",
     "symfony/ux-live-component": "^2.12",
+    "symfony/ux-turbo": "^2.24",
     "symfony/ux-twig-component": "^2.16",
     "symfony/validator": "7.0.*",
     "symfony/web-link": "7.0.*",
