+++ templates/_include/_sidePanel.html.twig
@@ -1,4 +1,7 @@
-<nav class="side-panel navbar navbar-expand-lg py-4 align-items-lg-start">
+<nav
+        {{ stimulus_controller('uiToggles') }}
+        class="side-panel navbar navbar-expand-lg py-4 align-items-lg-start"
+>
     <div class="container-fluid flex-lg-column">
         <img src="{{ asset('/images/mainlogo1.png') }}" alt="">
         <a class="navbar-brand m-0" href="{{ path('app_home') }}"><h1>Kopeck !</h1></a>
