+++ templates/base.html.twig
@@ -16,6 +16,11 @@
         {% endblock %}
         {{ pwa() }}
     {% endblock %}
+    <script>
+        document.documentElement.setAttribute(
+            'data-theme', localStorage.getItem('theme') || 'light'
+        );
+    </script>
 
 </head>
 <header>
