+++ templates/components/Form/SplitterForm.html.twig
@@ -51,11 +51,6 @@
                         <twig:ux:icon name="mdi:trash-can-outline" height="24" width="24" />
                     </button>
 
-
-
-
-
-
                 {% endif %}
 
             </div>
