+++ templates/expense/_form.html.twig
@@ -72,6 +72,6 @@
 
 
     <div class="centerBtn">
-        <button class="btn btn-primary">{{ button_label|default('Ajouter') }}</button>
+        <button class="btn btn-success valid-btn">{{ button_label|default('Ajouter') }}</button>
     </div>
 {{ form_end(form) }}
