+++ templates/expense/new.html.twig
@@ -4,16 +4,13 @@
 
 {% block body %}
 
-    <main class="container-md">
+    <main class="container-md expenseForm">
 
-        <div class="d-flex justify-content-between align-items-center my-3">
-
-            <h1>Ajouter une dépense</h1>
-
-            <a class="nav-link" href="{{ path('app_splitter_show', {'id': splitter.id}) }}" title="Retour">
-                <i class="bi bi-arrow-return-left"></i>
+        <div class="d-flex align-items-center gap-3 mb-4">
+            <a class="nav-link" href="{{ path('app_splitter_show', {'id': splitter.id}) }}" title="Retour Splitter">
+                <twig:ux:icon name="ep:back" height="30" width="30"/>
             </a>
-
+            <h1 class="text-center my-2">Ajouter une nouvelle dépense</h1>
         </div>
 
         {{ include('expense/_form.html.twig') }}
